Allow float values in default font size
I think it could be useful to allow users to set float values in DefaultFontSize, i.e. 10.5

+++ src/PhpWord/Settings.php
@@ -351,7 +351,7 @@ class Settings
      */
     public static function setDefaultFontSize($value)
     {
-        $value = intval($value);
+        $value = floatval($value);
         if ($value > 0) {
             self::$defaultFontSize = $value;
             return true;
